moved graphsimplification logic into simplify.rs
