+++ Cargo.toml
@@ -26,3 +26,4 @@ petgraph = "0.6.4"
 geo = "0.27.0"
 geojson = "0.24.1" 
 lru = "0.12.2"
+rtree_rs = "0.1.4"
